Refactor useDidUpdate
