+++ packages/blade/src/components/Collapsible/CollapsiblePanel.web.tsx
@@ -2,8 +2,8 @@ import type { ReactElement, ReactNode, TransitionEventHandler } from 'react';
 import { useRef } from 'react';
 import styled from 'styled-components';
 import { useCollapsible } from './CollapsibleContext';
+import { useDidUpdate } from './useDidUpdate';
 import { castWebType, makeMotionTime, makeSize } from '~utils';
-import { useDidUpdate } from '~src/hooks/useDidUpdate';
 import { Box } from '~components/Box';
 
 type CollapsiblePanelProps = {
